[refactor] 기존 외부 서버 이름을 external -> service 로 변경
