--- external/advertisement-external/src/main/java/woowa/team4/advertisementexternal/AdvertisementExternalApplication.java
@@ -1,13 +1,0 @@
-package woowa.team4.advertisementexternal;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AdvertisementExternalApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AdvertisementExternalApplication.class, args);
-    }
-
-}
