--- external/advertisement-external/src/test/java/woowa/team4/advertisementexternal/AdvertisementExternalApplicationTests.java
@@ -1,13 +1,0 @@
-package woowa.team4.advertisementexternal;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AdvertisementExternalApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
