--- external/coupon-external/src/main/java/woowa/team4/couponexternal/CouponExternalApplication.java
@@ -1,13 +1,0 @@
-package woowa.team4.couponexternal;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CouponExternalApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CouponExternalApplication.class, args);
-    }
-
-}
