--- external/coupon-external/src/test/java/woowa/team4/couponexternal/CouponExternalApplicationTests.java
@@ -1,13 +1,0 @@
-package woowa.team4.couponexternal;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CouponExternalApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
