--- external/delivery-time-external/src/main/java/woowa/team4/deliverytimeexternal/DeliveryTimeExternalApplication.java
@@ -1,13 +1,0 @@
-package woowa.team4.deliverytimeexternal;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DeliveryTimeExternalApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DeliveryTimeExternalApplication.class, args);
-    }
-
-}
