--- external/delivery-time-external/src/test/java/woowa/team4/deliverytimeexternal/DeliveryTimeExternalApplicationTests.java
@@ -1,13 +1,0 @@
-package woowa.team4.deliverytimeexternal;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DeliveryTimeExternalApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
