+++ service/coupon-service/src/main/java/woowa/team4/bff/CouponServiceApplication.java
@@ -1,0 +1,13 @@
+package woowa.team4.bff;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CouponServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CouponServiceApplication.class, args);
+    }
+
+}
