+++ service/delivery-time-service/src/main/java/woowa/team4/bff/DeliveryTimeServiceApplication.java
@@ -1,0 +1,13 @@
+package woowa.team4.bff.deliverytimeservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DeliveryTimeServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(DeliveryTimeServiceApplication.class, args);
+    }
+
+}
